continous improvement

+++ Fracture/src/Rendering/Mesh.cpp
@@ -60,7 +60,7 @@ void Fracture::Mesh::setupMesh()
 
     glBindVertexArray(0);
 
-    IndexCount = m_indices.size();
+    IndexCount = (unsigned int)m_indices.size();
 
 
 }
